IEBH-449: Refactor debug fixture usage
