+++ tests/conftest.py
@@ -6,5 +6,6 @@
 
 pytest_plugins = [
     'tests.fixtures.context',
+    'tests.fixtures.debug',
     'tests.fixtures.fake',
 ]
