+++ tests/test_file_explorer.py
@@ -8,7 +8,6 @@ import hashlib
 import io
 import os
 import re
-import time as tm
 import zipfile
 from collections.abc import Generator
 from contextlib import contextmanager
@@ -28,20 +27,6 @@ from playwright.sync_api import expect
 from pydantic import BaseModel
 
 from tests.fixtures.fake import Fake
-
-
-class Debug:
-    def __init__(self, output_dir: Path) -> None:
-        self.output_dir = output_dir
-
-    def capture_screenshot(self, page: Page) -> None:
-        page.screenshot(path=self.output_dir / f'debug-screenshot-{int(tm.time())}.png')
-
-
-@pytest.fixture
-def debug(pytestconfig: pytest.Config) -> Debug:
-    output_dir = Path(pytestconfig.getoption('--output')).absolute()
-    return Debug(output_dir)
 
 
 class FileAttribute(BaseModel):
@@ -74,10 +59,9 @@ class File(BaseModel):
 
 
 class FileExplorer:
-    def __init__(self, page: Page, project_code: str, debug: Debug | None = None) -> None:
+    def __init__(self, page: Page, project_code: str) -> None:
         self.page = page
         self.project_code = project_code
-        self.debug = debug
 
     def open(self) -> Self:
         with self.page.expect_response(lambda r: 'v1/files/meta?' in r.url and 'order_by=created_time' in r.url):
@@ -103,8 +87,6 @@ class FileExplorer:
             self.locate_row(name).get_by_role('checkbox').check()
 
         with self.page.expect_download() as download_info:
-            if self.debug:
-                self.debug.capture_screenshot(self.page)
             self.page.get_by_role('button', name='cloud-download Download', exact=True).click()
 
         return download_info.value
@@ -274,10 +256,10 @@ class FileExplorer:
             yield
 
 
-def test_file_upload_and_download(admin_page: Page, project_code: str, working_path: Path, debug: Debug) -> None:
+def test_file_upload_and_download(admin_page: Page, project_code: str, working_path: Path) -> None:
     """Test that a file can be uploaded and then downloaded successfully."""
 
-    file_explorer = FileExplorer(admin_page, project_code, debug)
+    file_explorer = FileExplorer(admin_page, project_code)
     file_explorer.open().create_folders_and_navigate_to(working_path / 'file-upload')
 
     file = File.generate()
@@ -334,12 +316,10 @@ def test_file_upload_with_attributes(admin_page: Page, project_code: str, workin
     )
 
 
-def test_folder_upload_and_download(
-    admin_page: Page, project_code: str, working_path: Path, tmp_path: Path, debug: Debug
-) -> None:
+def test_folder_upload_and_download(admin_page: Page, project_code: str, working_path: Path, tmp_path: Path) -> None:
     """Test that a folder can be uploaded and then downloaded successfully."""
 
-    file_explorer = FileExplorer(admin_page, project_code, debug)
+    file_explorer = FileExplorer(admin_page, project_code)
     file_explorer.open().create_folders_and_navigate_to(working_path / 'folder-upload')
 
     folder_name = f'e2e-test-{os.urandom(5).hex()}'
@@ -361,12 +341,10 @@ def test_folder_upload_and_download(
 
 
 @pytest.mark.skip(reason='Resumable upload has a bug that needs to be fixed')
-def test_file_resumable_upload_and_download(
-    admin_page: Page, project_code: str, working_path: Path, debug: Debug
-) -> None:
+def test_file_resumable_upload_and_download(admin_page: Page, project_code: str, working_path: Path) -> None:
     """Test that an interrupted file upload can be resumed and then successfully downloaded."""
 
-    file_explorer = FileExplorer(admin_page, project_code, debug)
+    file_explorer = FileExplorer(admin_page, project_code)
     file_explorer.open().create_folders_and_navigate_to(working_path / 'file-resume-upload')
 
     file = File.generate(size_kb=4096)
